chore: address review feedback
